Move disassemble tests to compiletest
