+++ crates/spirv-builder/Cargo.toml
@@ -18,7 +18,3 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 # See comment in lib.rs invoke_rustc for why this is here
 rustc_codegen_spirv = { path = "../rustc_codegen_spirv", default-features = false }
-
-[dev-dependencies]
-lazy_static = "1.4"
-pretty_assertions = "0.7"
