+++ tests/ui/dis/asm.rs
@@ -1,0 +1,19 @@
+// build-pass
+// compile-flags: -C llvm-args=--disassemble-fn=asm::asm
+
+use spirv_std as _;
+
+fn asm() {
+    unsafe {
+        asm!(
+            "%int = OpTypeInt 32 0",
+            "%scope = OpConstant %int 2",
+            "%semantics = OpConstant %int 16",
+            "OpMemoryBarrier %scope %semantics",
+        );
+    }
+}
+#[spirv(fragment)]
+pub fn main() {
+    asm();
+}
