+++ tests/ui/dis/asm.stderr
@@ -1,0 +1,5 @@
+%1 = OpFunction  %2  None %3
+%4 = OpLabel
+OpMemoryBarrier %5 %6
+OpReturn
+OpFunctionEnd
