+++ tests/ui/dis/asm_add_two_ints.stderr
@@ -1,0 +1,7 @@
+%1 = OpFunction  %2  None %3
+%4 = OpFunctionParameter  %2
+%5 = OpFunctionParameter  %2
+%6 = OpLabel
+%7 = OpIAdd  %2  %4 %5
+OpReturnValue %7
+OpFunctionEnd
