+++ tests/ui/dis/asm_op_decorate.rs
@@ -1,0 +1,37 @@
+// build-pass
+// compile-flags: -C llvm-args=--disassemble-globals
+
+use spirv_std as _;
+
+fn add_decorate() {
+    unsafe {
+        let offset = 1u32;
+        asm!(
+            "OpExtension \"SPV_EXT_descriptor_indexing\"",
+            "OpCapability RuntimeDescriptorArray",
+            "OpDecorate %image_2d_var DescriptorSet 0",
+            "OpDecorate %image_2d_var Binding 0",
+            "%uint                  = OpTypeInt 32 0",
+            "%float                 = OpTypeFloat 32",
+            "%uint_0                = OpConstant %uint 0",
+            "%image_2d              = OpTypeImage %float Dim2D 0 0 0 1 Unknown",
+            "%sampled_image_2d      = OpTypeSampledImage %image_2d",
+            "%image_array           = OpTypeRuntimeArray %sampled_image_2d",
+            // NOTE(eddyb) `Generic` is used here because it's the placeholder
+            // for storage class inference - both of the two `OpTypePointer`
+            // types below should end up inferring to `UniformConstant`.
+            "%ptr_image_array       = OpTypePointer Generic %image_array",
+            "%image_2d_var          = OpVariable %ptr_image_array UniformConstant",
+            "%ptr_sampled_image_2d  = OpTypePointer Generic %sampled_image_2d",
+            "", // ^^ type preamble
+            "%offset                = OpLoad _ {0}",
+            "%24                    = OpAccessChain %ptr_sampled_image_2d %image_2d_var %offset",
+            "%25                    = OpLoad %sampled_image_2d %24",
+            in(reg) &offset,
+        );
+    }
+}
+#[spirv(fragment)]
+pub fn main() {
+    add_decorate();
+}
