+++ tests/ui/dis/asm_op_decorate.stderr
@@ -1,0 +1,22 @@
+OpCapability Shader
+OpCapability RuntimeDescriptorArray
+OpExtension "SPV_EXT_descriptor_indexing"
+OpMemoryModel Logical Simple
+OpEntryPoint Fragment %1 "main"
+OpExecutionMode %1 OriginUpperLeft
+OpName %2 "asm_op_decorate::add_decorate"
+OpName %3 "asm_op_decorate::main"
+OpDecorate %4 ArrayStride 4
+OpDecorate %5 DescriptorSet 0
+OpDecorate %5 Binding 0
+%6 = OpTypeVoid
+%7 = OpTypeFunction %6
+%8 = OpTypeInt 32 0
+%9 = OpConstant  %8  1
+%10 = OpTypeFloat 32
+%11 = OpTypeImage %10 2D 0 0 0 1 Unknown
+%12 = OpTypeSampledImage %11
+%4 = OpTypeRuntimeArray %12
+%13 = OpTypePointer UniformConstant %4
+%5 = OpVariable  %13  UniformConstant
+%14 = OpTypePointer UniformConstant %12
