+++ tests/ui/dis/complex_image_sample_inst.stderr
@@ -1,0 +1,12 @@
+%1 = OpFunction  %2  None %3
+%4 = OpFunctionParameter  %2
+%5 = OpFunctionParameter  %6
+%7 = OpFunctionParameter  %6
+%8 = OpFunctionParameter  %9
+%10 = OpFunctionParameter  %9
+%11 = OpLabel
+%12 = OpAccessChain  %13  %14 %15
+%16 = OpLoad  %17  %12
+%18 = OpImageSampleProjExplicitLod  %2  %16 %4 Grad|ConstOffset %5 %7 %19
+OpReturnValue %18
+OpFunctionEnd
