+++ tests/ui/dis/custom_entry_point.stderr
@@ -1,0 +1,7 @@
+OpCapability Shader
+OpMemoryModel Logical Simple
+OpEntryPoint Fragment %1 "hello_world"
+OpExecutionMode %1 OriginUpperLeft
+OpName %2 "custom_entry_point::main"
+%3 = OpTypeVoid
+%4 = OpTypeFunction %3
