+++ tests/ui/dis/ptr_copy.stderr
@@ -1,0 +1,7 @@
+%1 = OpFunction  %2  None %3
+%4 = OpFunctionParameter  %5
+%6 = OpFunctionParameter  %5
+%7 = OpLabel
+OpCopyMemory %6 %4
+OpReturn
+OpFunctionEnd
