+++ tests/ui/dis/ptr_read_method.stderr
@@ -1,0 +1,11 @@
+%1 = OpFunction  %2  None %3
+%4 = OpFunctionParameter  %5
+%6 = OpFunctionParameter  %5
+%7 = OpLabel
+%8 = OpVariable  %5  Function
+OpStore %8 %9
+OpCopyMemory %8 %4
+%10 = OpLoad  %11  %8
+OpStore %6 %10
+OpReturn
+OpFunctionEnd
