+++ tests/ui/dis/ptr_write.stderr
@@ -1,0 +1,10 @@
+%1 = OpFunction  %2  None %3
+%4 = OpFunctionParameter  %5
+%6 = OpFunctionParameter  %5
+%7 = OpLabel
+%8 = OpVariable  %5  Function
+%9 = OpLoad  %10  %4
+OpStore %8 %9
+OpCopyMemory %6 %8
+OpReturn
+OpFunctionEnd
